requirements.txt
pyspark==3.5.0
pyyaml==6.0.3
delta-spark==3.2.0
